Crate coupe

A mesh partitioning library that implements multithreaded, composable geometric algorithms.

Crate Layout

Coupe exposes a Partition trait, which is in turn implemented by algorithms. See its documentation for more details. The trait is generic around its input, which means algorithms can partition different type of collections (e.g. 2D and 3D meshes).

Available algorithms

Partitioner algorithms

Space filling curves:

Z-curve

Hilbert curve

Recursive Coordinate Bisection

Recursive Inertial Bisection

Multi jagged

Number partitioning:

Greedy

Karmarkar-Karp and its complete version

Partition improving algorithms

K-means

Number partitioning:

VN-Best

VN-First

Fiduccia-Mattheyses

Kernighan-Lin

Re-exports

pub use nalgebra;

pub use rayon;

pub use sprs;

Modules

imbalance

num_traits

topology

Utilities to handle topologic concepts and metrics related to mesh

Structs

BoundingBox

Axis-aligned bounding box.

CompleteKarmarkarKarp

Complete Karmarkar-Karp algorithm

FiducciaMattheyses

FiducciaMattheyses

FmMetadata

Diagnostic data for a Fiduccia-Mattheyses run.

GraphGrowth

Graph Growth algorithm

Greedy

Greedy number partitioning algorithm

HilbertCurve

Hilbert space-filling curve algorithm

KMeans

K-means algorithm

KarmarkarKarp

Karmarkar-Karp algorithm

KernighanLin

KernighanLin algorithm

MultiJagged

Multi-Jagged algorithm

Random

Map elements to parts randomly.

Rcb

Recursive Coordinate Bisection algorithm

Real

A floating-point value that cannot be NAN nor infinity.

Rib

Recursive Inertial Bisection algorithm

VnBest

Steepest descent Vector-of-Numbers algorithm

VnFirst

Descent Vector-of-Numbers algorithm

ZCurve

Z space-filling curve algorithm

Enums

Error

Common errors thrown by algorithms.

HilbertCurveError

Traits

CkkWeight

Trait alias for values accepted as weights by CompleteKarmarkarKarp.

FmWeight

Trait alias for values accepted as weights by FiducciaMattheyses.

GreedyWeight

Trait alias for values accepted as weights by Greedy.

KkWeight

Trait alias for values accepted as weights by KarmarkarKarp.

Partition

The Partition trait allows for partitioning data.

RcbWeight

Trait alias for values accepted as weights by Rcb and Rib.

VnBestWeight

Trait alias for values accepted as weights by VnBest.

VnFirstWeight

Trait alias for values accepted as weights by VnFirst.

Type Definitions

Point2D

Point3D

PointND