Function coupe::topology::adjacency_matrix

pub fn adjacency_matrix(
    conn: CsMatView<'_, u32>, 
    num_common_nodes: u32
) -> CsMat<f64>

Build an adjacency matrix from a mesh connectivity.

Inputs

conn: the connectivity of a mesh. It’s a sparse matrix in which the (i, j) entry is equal to 1 if and only if nodes i and j are linked.

num_common_nodes: defines how the adjacency matrix is build. Two elements of the mesh are linked in the adjacency graph is they have exactly num_common_nodes nodes in common.

Output

A sparse matrix in which the enty (i, j) is non-zero if and only if the mesh elements i and j are linked in the mesh graph.

If the entry (i, j) is non-zero, then its value is the weight of the edge between i and j (default to 1.0).